Metadata List to XML Converter

This function converts a list of metadata to XML

Usage

build_metadata_xml_from_list(
  input_data,
  metatype = NULL,
  root_name = NULL,
  ns = c(character(0)),
  root = NULL
)

Arguments

input_data

XML document serving as the basis upon which to add the list

metatype

a character indicating the element name of each record in the list

root_name

character; the name of the root node if created

ns

named vector; a collection of character strings indicating the namespace definitions of the root node if created

root

xmlNode; an XML node to be used as the root

Value

xmlNode; an XML node with the input data added as needed for the Metadata API and its objects.

Note

This function is meant to be used internally. Only use when debugging.