Pulls out a tibble of record info from an XML node

This function accepts an xml_nodeset and searches for all './/records' in the document to format into a single tidy tbl_df.

Usage

extract_records_from_xml_nodeset(
  nodeset,
  object_name_append = FALSE,
  object_name_as_col = FALSE
)

Arguments

nodeset

xml_nodeset; nodeset to have records extracted into a tbl_df

object_name_append

logical; whether to include the object type (e.g. Account or Contact) as part of the column names (e.g. Account.Name).

object_name_as_col

logical; whether to include the object type (e.g. Account or Contact) as a new column.

Value

tbl_df parsed from the supplied xml_nodeset

Note

This function is meant to be used internally. Only use when debugging.