Bulk Job Records URL Generator

Bulk Job Records URL Generator

Usage

make_bulk_job_records_url(
  job_id,
  record_type = c("successfulResults", "failedResults", "unprocessedRecords"),
  api_type = c("Bulk 2.0")
)

Arguments

job_id

character; the Salesforce Id assigned to a submitted job as returned by sf_create_job_bulk. It will start with "750".

record_type

character; one of 'successfulResults', 'failedResults', or 'unprocessedRecords' indicating the type of records to retrieve.

api_type

character; a string indicating which Bulk API to execute the call against.

Value

character; a complete URL (as a string) that will be used to send subsequent Bulk API calls to. This URL is specific to your instance and the API version being used.

Note

This function is meant to be used internally. Only use when debugging.