Query URL Generator

Query URL Generator

Usage

make_query_url(soql, queryall, next_records_url)

Arguments

soql

character; a string defining a SOQL query (e.g. "SELECT Id, Name FROM Account").

queryall

logical; indicating if the query recordset should include records that have been deleted because of a merge or delete. Setting this argument to TRUE will also return information about archived Task and Event records. It is available in API versions 29.0 and later.

next_records_url

character; a string returned by a Salesforce query from where to find subsequent records returned by a paginated query.

Value

character; a complete URL (as a string) to send a request to in order to retrieve queried records.

Note

This function is meant to be used internally. Only use when debugging.