salesforcer's backwards compatible version of rforcecom.create

salesforcer's backwards compatible version of rforcecom.create

rforcecom.create(session, objectName, fields)

Arguments

session
list; a list containing "sessionID", "instanceURL", and " apiVersion" as returned by RForcecom::rforcecom.login(). This argument is ignored in all backward compatible calls because the authorization credentials are stored in an environment internal to the salesforcer package, so it is no longer necessary to pass the session in each function call.

objectName
character; the name of the Salesforce object that the function is operating against (e.g. "Account", "Contact", "CustomObject__c")

fields
Field names and values. (ex: Name="CompanyName", Phone="000-000-000" )

Value

data.frame containing the id and success indicator of the record creation process