The salesforcer backwards compatible version of rforcecom.retrieve

rforcecom.retrieve(
  session,
  objectName,
  fields,
  limit = NULL,
  id = NULL,
  offset = NULL,
  order = NULL,
  inverse = NULL,
  nullsLast = NULL
)

Arguments

session
list; a list containing "sessionID", "instanceURL", and "apiVersion" as returned by rforcecom.login. This argument is ignored in all backward compatible calls because the authorization credentials are stored in an environment internal to the salesforcer package, so it is no longer necessary to pass the session in each function call.

objectName
character; the name of the Salesforce object that the function is operating against (e.g. "Account", "Contact", "CustomObject__c").

fields
A List of field names. (ex: c("Id", "Name", "Industry", "AnnualRevenue)"))

limit
Number of the records to retrieve. (ex: 5)

id
Record ID to retrieve. (ex: "999x000000xxxxxZZZ")

offset
Specifies the starting row offset. (ex: "100")

order
A list for controling the order of query results. (ex: "c("Industry","Name")")

inverse
If it is TRUE, the results are ordered in descending order. This parameter works when order parameter has been set. (Default: FALSE)

nullsLast
If it is TRUE, null records list in last. If not null records list in first. This parameter works when order parameter has been set. (Default: FALSE)