Create Bulk API Job

This function initializes a Job in the Salesforce Bulk API

sf_bulk_create_job(operation = c("insert", "delete", "upsert", "update",
  "hardDelete", "query"), object, content_type = c("CSV", "ZIP_CSV",
  "ZIP_XML", "ZIP_JSON"), external_id_fieldname = NULL,
  concurrency_mode = c("Parallel", "Serial"), line_ending = NULL,
  column_delimiter = NULL, verbose = FALSE)

Arguments

operation
character; a string defining the type of operation being performed (e.g. "insert", "update", "upsert", "delete")

object
character; the name of the Salesforcer object that the function is operating against (e.g. "Account", "Contact", "CustomObject__c")

content_type
a character string being one of 'CSV','ZIP_CSV','ZIP_XML', or 'ZIP_JSON'

external_id_fieldname
character; string identifying a custom field on the object that has been set as an "External ID" field. This field is used to reference objects during upserts to determine if the record already exists in Salesforce or not.

concurrency_mode
a character string either "Parallel" or "Serial" that specifies whether batches should be completed sequentially or in parallel. Use "Serial" only if Lock contentions persist with in "Parallel" mode.

line_ending
a character string indicating the The line ending used for CSV job data, marking the end of a data row. The default is "LF", but could be "CRLF" for Windows created files.

column_delimiter
a character string indicating the column delimiter used for CSV job data. The default value is COMMA. Valid values are: "BACKQUOTE", "CARET", "COMMA", "PIPE", "SEMICOLON", and "TAB".

verbose
logical; do you want informative messages?

Value

A list parameters defining the created job, including id

References

https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/

Examples

# NOT RUN {
# insert into Account
job_info <- sf_bulk_create_job(operation='insert', object='Account')

# delete from Account
job_info <- sf_bulk_create_job(operation='delete', object='Account')

# update into Account
job_info <- sf_bulk_create_job(operation='update', object='Account')

# upsert into Account
job_info <- sf_bulk_create_job(operation='upsert',
                               externalIdFieldName='My_External_Id__c',
                               object='Account')

# insert attachments
job_info <- sf_bulk_create_job(operation='insert', object='Attachment')

# query leads
job_info <- sf_bulk_create_job(operation='query', object='Lead')
# }