Get the results of an existing dashboard

This function allows for pulling specific data from a dashboard. There is a convenience function (sf_get_dashboard_data) to get the dashboard data in a tabular format returned as a tbl_df.

sf_dashboard_results(
  dashboard_id,
  running_user = NULL,
  dashboard_filters = c(character(0))
)

Arguments

dashboard_id
character; the Salesforce Id assigned to a created dashboard. It will start with "01Z".

running_user
character; the Salesforce Id that should be assigned as the runner of the job. This should be from a User record (i.e. the Id will start with "005"). Note that this will throw an error if the User is not allowed to change the running User, or if the selected running User is invalid.

dashboard_filters
character; Dashboard results are always unfiltered, unless you have specified filter parameters in your request. Use this argument to include up to three optional filter Ids. You can obtain the list of defined filter Ids from the dashboard metadata using sf_dashboard_describe.

Value

tbl_df