Describe a report type

Return metadata about a report type.

Usage

sf_describe_report_type(report_type, verbose = FALSE)

Arguments

report_type

character; a character representing the type of report to retrieve the metadata information on. A list of valid report types that can be described using this function will be available in the reportTypes.type column of results returned sf_list_report_types. (e.g. AccountList, AccountContactRole, OpportunityHistory, etc.)

verbose

logical; an indicator of whether to print additional detail for each API call, which is useful for debugging. More specifically, when set to TRUE the URL, header, and body will be printed for each request, along with additional diagnostic information where available.

Value

list containing up to 4 properties that describe the report:

attributes

Report type along with the URL to retrieve common objects and joined metadata.

reportMetadata

Unique identifiers for groupings and summaries.

reportTypeMetadata

Fields in each section of a report type plus filter information for those fields.

reportExtendedMetadata

Additional information about summaries and groupings.

Salesforce Documentation

Documentation

See also

Examples

if (FALSE) { # \dontrun{
reports <- sf_list_report_types()
unique_report_types <- reports %>% distinct(reportTypes.type)

# first unique report type
unique_report_types[[1,1]]

# describe that report type
described_report <- sf_describe_report_type(unique_report_types[[1,1]])
} # }