Get Deleted Records from a Timeframe

Retrieves the list of individual records that have been deleted within the given timespan for the specified object.

Usage

sf_get_deleted(object_name, start, end, verbose = FALSE)

Arguments

object_name

character; the name of the Salesforce object that the function is operating against (e.g. "Account", "Contact", "CustomObject__c").

start

date or datetime; starting datetime of the timespan for which to retrieve the data.

end

date or datetime; ending datetime of the timespan for which to retrieve the data.

verbose

logical; an indicator of whether to print additional detail for each API call, which is useful for debugging. More specifically, when set to TRUE the URL, header, and body will be printed for each request, along with additional diagnostic information where available.

Note

This API ignores the seconds portion of the supplied datetime values.

Examples

if (FALSE) { # \dontrun{
# get all deleted Contact records from midnight until now
deleted_recs <- sf_get_deleted("Contact", Sys.Date(), Sys.time())
} # }