Validate the input for an operation

Validate the input for an operation

Usage

sf_input_data_validation(input_data, operation = "")

Arguments

input_data

named vector, matrix, data.frame, or tbl_df; data can be coerced into a data.frame. If performing an update, upsert, or delete operation, then one column or field must be the Id of the record to modify or delete.

operation

character; a string defining the type of operation being performed (e.g. "insert", "update", "upsert", "delete", "hardDelete", "query", "queryall").

Value

the input data validated and formatted according to the specified operation. This allows more flexibility for the user while ensuring that all inputs are formatted as expected by the target APIs and operations.

Note

This function is meant to be used internally. Only use when debugging.