List the Limits for an API

Lists information about limits in your org.

sf_list_api_limits()

Value

list

Note

This resource is available in REST API version 29.0 and later for API users with the View Setup and Configuration permission. The resource returns these limits:

Daily API calls

Daily asynchronous Apex method executions (batch Apex, future methods, queueable Apex, and scheduled Apex)

Daily Bulk API calls

Daily Streaming API events (API version 36.0 and earlier)

Daily durable Streaming API events (API version 37.0 and later)

Streaming API concurrent clients (API version 36.0 and earlier)

Durable Streaming API concurrent clients (API version 37.0 and later)

Daily generic streaming events (API version 36.0 and earlier)

Daily durable generic streaming events (API version 37.0 and later)

Daily number of mass emails that are sent to external email addresses by using Apex or APIs

Daily number of single emails that are sent to external email addresses by using Apex or APIs

Concurrent REST API requests for results of asynchronous report runs

Concurrent synchronous report runs via REST API

Hourly asynchronous report runs via REST API

Hourly synchronous report runs via REST API

Hourly dashboard refreshes via REST API

Hourly REST API requests for dashboard results

Hourly dashboard status requests via REST API

Daily workflow emails

Hourly workflow time triggers

Hourly OData callouts

Daily and active scratch org counts

Data storage (MB)

File storage (MB)

Examples

# NOT RUN {
sf_list_api_limits()
# }