Perform SOQL Query

Executes a query against the specified object and returns data that matches the specified criteria.

sf_query(soql, object = NULL, queryall = FALSE, batch_size = 1000,
  api_type = c("REST", "SOAP", "Bulk", "Async"), next_records_url = NULL,
  ..., verbose = FALSE)

Arguments

soql
character; a string defining a SOQL query (e.g. "SELECT Id, Name FROM Account")

object
character; the name of the Salesforcer object that the function is operating against (e.g. "Account", "Contact", "CustomObject__c")

queryall
logical; indicating if the query recordset should include deleted and archived records (available only when querying Task and Event records)

batch_size
numeric; a number between 200 and 2000 indicating the number of records per page that are returned. Speed benchmarks should be done to better understand the speed implications of choosing either endpoint.

api_type
character; one of "REST", "SOAP", "Bulk", or "Async" indicating which API to use when making the request

next_records_url
character (leave as NULL); a string used internally by the function to paginate through to more records until complete

...
Other arguments passed on to sf_bulk_query.

verbose
logical; do you want informative messages?

Value

tibble

Note

Bulk API query doesn't support the following SOQL:

COUNT

ROLLUP

SUM

GROUP BY CUBE

OFFSET

Nested SOQL queries

Relationship fields

Additionally, Bulk API can't access or query compound address or compound geolocation fields.

References

https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/

Examples

# NOT RUN {
sf_query("SELECT Id, Account.Name, Email FROM Contact LIMIT 10")
sf_query("SELECT Id, Email FROM Contact LIMIT 10", verbose = TRUE)
# }