Rename Metadata Elements in Salesforce

This function takes an old and new name for a metadata element in Salesforce and applies the new name

sf_rename_metadata(metadata_type, old_fullname, new_fullname, verbose = FALSE)

Arguments

metadata_type
character; string on what type of metadata to create.

old_fullname
character; string corresponding to the fullName of the element you would like to rename

new_fullname
character; string corresponding to the new fullName you would like to apply the targeted element

verbose
logical; an indicator of whether to print additional detail for each API call, which is useful for debugging. More specifically, when set to TRUE the URL, header, and body will be printed for each request, along with additional diagnostic information where available.

Value

A data.frame containing the creation result for each submitted metadata component

References

https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/

Examples

if (FALSE) {
renamed_custom_object <- sf_rename_metadata(metadata_type = 'CustomObject',
                                            old_fullname = 'Custom_Account32__c',
                                            new_fullname = 'Custom_Account99__c')
}