Set a sticky dashboard filter

Set a default filter value which gets applied to a dashboard when you open it. The default filter value you specify only applies to you (other people won’t see it when they open the dashboard). If you change the filter value while viewing the dashboard, then the filter value you set in the user interface overwrites the value you set via the API. To set sticky filters for a dashboard, canUseStickyFilter must equal true. Saves any dashboard filters set in the request so that they’re also set the next time you open the dashboard. NOTE: You can only set dashboard filters for yourself, not for other users.

Usage

sf_set_dashboard_sticky_filter(
  dashboard_id,
  dashboard_filters = c(character(0))
)

Arguments

dashboard_id

character; the Salesforce Id assigned to a created dashboard. It will start with "01Z".

dashboard_filters

character; Dashboard results are always unfiltered, unless you have specified filter parameters in your request. Use this argument to include up to three optional filter Ids. You can obtain the list of defined filter Ids from the dashboard metadata using sf_describe_dashboard.

Value

list